fix: apply new UI layout based on mockups
