+++ src/Languages/appTranslationTablesViews/all-views/en.json
@@ -93,12 +93,17 @@
   "INVALID_OLD_PASSWORD": "Incorrect old password",
   "INVALID_PASSWORD": "Invalid password.",
   "INVESTIGATOR": "Investigator",
+  "IPS_DESCRIPTION_PART_1": "An International Patient Summary (IPS) is a set of clinical data that provides an overview of your health data. It lists information such as your medical issues, medications, allergies, and more.",
+  "IPS_DESCRIPTION_PART_2": "You can use your IPS to securely share a snapshot of your health information with clinical staff who might not otherwise have access to your medical history, for example, when visiting a new doctor or clinic.",
   "IPS_FULL": "International Patient Summary",
-  "IPS_PROTOTYPE": "Prototype",
+  "IPS_OPEN_EXTERNAL": "Open in an external browser",
+  "IPS_PREVIEW_OR_SHARE": "Preview or Share IPS",
   "IPS_QR_CODE_ALT": "QR code to share the patient's International Patient Summary",
-  "IPS_QR_CODE_DESCRIPTION": "The following QR code and associated link provide direct viewing access to your patient summary. Protect them as you would a password, or your printed medical file.",
-  "IPS_QR_CODE_LINK": "QR code link:",
-  "IPS_SWITCH_VIEWER": "Switch viewer",
+  "IPS_SHARE_DESCRIPTION_PART_1": "The IPS displayed in the Preview section below can be shared via the following QR code or external link, which will",
+  "IPS_SHARE_DESCRIPTION_PART_2": "expire in 1 hour",
+  "IPS_SHARE_DESCRIPTION_PART_3": "If you leave this page and come back, a new code will be generated.",
+  "IPS_SHARE_DESCRIPTION_PART_4": "This QR code and link provide direct viewing access to your patient summary. Protect them as you would a password, or your printed medical file.",
+  "IPS_SHORT": "IPS",
   "IPS_VIEWER_ALT": "IPS viewer",
   "LABELS": "Labels",
   "LAB_BY_DATE": "By Date",
@@ -167,6 +172,7 @@
   "PHONE_NUMBER": "Phone Number",
   "PLOT_VIEW": "Plot View",
   "PREFERENCES": "Preferences",
+  "PREVIEW": "Preview",
   "QUEBEC_MEDICAL_SCHEDULER": "Québec Medical Appointment Scheduler",
   "QUESTION": "Question",
   "QUESTIONNAIRES": "Questionnaires",
