+++ src/Languages/appTranslationTablesViews/all-views/fr.json
@@ -93,12 +93,17 @@
   "INVALID_OLD_PASSWORD": "Mot de passe actuel invalide",
   "INVALID_PASSWORD": "Mot de passe invalide.",
   "INVESTIGATOR": "Chercheur(euse)",
+  "IPS_DESCRIPTION_PART_1": "Le résumé international du dossier médical du patient (IPS) est un ensemble de données cliniques qui fournit un aperçu de vos données de santé. Il contient des informations sur vos problèmes de santé, vos médicaments, vos allergies, etc.",
+  "IPS_DESCRIPTION_PART_2": "Vous pouvez utiliser votre IPS pour partager de façon sécurisée un portrait de vos informations de santé avec du personnel clinique qui n'aurait pas autrement accès à vos antécédents médicaux. Par exemple, vos pourriez vous en servir lors d'une visite chez un nouveau médecin ou dans une nouvelle clinique.",
   "IPS_FULL": "Résumé international du dossier médical du patient",
-  "IPS_PROTOTYPE": "Prototype",
+  "IPS_OPEN_EXTERNAL": "Ouvrir dans un navigateur externe",
+  "IPS_PREVIEW_OR_SHARE": "Visualiser ou partager l'IPS",
   "IPS_QR_CODE_ALT": "Code QR pour partager le résumé international du dossier médical du patient",
-  "IPS_QR_CODE_DESCRIPTION": "Le code QR suivant et l'hyperlien représenté par celui-ci permettent de visualiser directement votre résumé de dossier patient. Protégez-les comme vous le feriez avec un mot de passe ou une copie imprimée de votre dossier médical.",
-  "IPS_QR_CODE_LINK": "Hyperlien du code QR :",
-  "IPS_SWITCH_VIEWER": "Changer de vue",
+  "IPS_SHARE_DESCRIPTION_PART_1": "Le résumé IPS affiché dans la section Aperçu ci-dessous peut être partagé avec le code QR ou le lien externe suivants, qui",
+  "IPS_SHARE_DESCRIPTION_PART_2": "expirent dans 1 heure",
+  "IPS_SHARE_DESCRIPTION_PART_3": "Si vous quittez cette page et y revenez, un nouveau code sera créé.",
+  "IPS_SHARE_DESCRIPTION_PART_4": "Le code QR et le lien suivants permettent de visualiser directement votre résumé de dossier patient. Protégez-les comme vous le feriez avec un mot de passe ou une copie imprimée de votre dossier médical.",
+  "IPS_SHORT": "Résumé IPS",
   "IPS_VIEWER_ALT": "Visualiseur du résumé de dossier patient",
   "LABELS": "Étiquettes",
   "LAB_BY_DATE": "Date",
@@ -167,6 +172,7 @@
   "PHONE_NUMBER": "Téléphone",
   "PLOT_VIEW": "Vue Graphique",
   "PREFERENCES": "Préférences",
+  "PREVIEW": "Aperçu",
   "QUEBEC_MEDICAL_SCHEDULER": "Rendez-vous santé Québec",
   "QUESTION": "Question",
   "QUESTIONNAIRES": "Questionnaires",
