+++ src/css/views/ips.view.css
@@ -4,47 +4,62 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-.ips-iframe {
-    height: 92%;
+.ips--button-external {
+    background-color: white;
+    color: #6d6d6d;
+    font-weight: normal;
+    margin: 10px auto;
+    width: 95%;
+}
+
+.ips--button-preview-share {
+    background-color: #ddebff;
+    color: #1f1f21;
+    font-weight: normal;
+    margin-top: 30px;
+}
+
+.ips--description {
+    padding: 30px;
+}
+
+.ips--description-sharing {
+    font-size: 0.9em;
+}
+
+.ips--description-sharing p:last-child {
+    margin-bottom: 5px;
+}
+
+.ips--iframe {
+    height: 1000px;
     width: 100%;
 }
 
-.ips-outer-panel {
+.ips--inner-panel {
+    margin: 20px;
+}
+
+.ips--outer-panel {
     margin: 10px;
 }
 
-.ips-outer-panel .panel-heading {
+.ips--outer-panel .panel-heading {
     background-color: #e7f1ff;
 }
 
-.ips-outer-panel .panel-title {
+.ips--outer-panel .panel-title {
     font-size: 1em;
 }
 
-.ips-prototype-badge {
-    background-color: #e3e3e3;
-    color: #6f6f6f;
-}
-
-.ips-qr-code {
+.ips--qr-code {
     display: block;
     margin: auto;
+    max-width: fit-content;
+    width: 90%;
 }
 
-.ips-link {
-    color: #337ab7;
-    margin: 0 10px 20px;
-    padding: 10px;
-}
-
-.ips-link-label {
-    margin: 10px 10px 0;
-    padding: 10px;
-}
-
-.ips-viewer-badge {
-    background-color: #b9dce8;
-    color: #475f78;
-    float: right;
-    margin: 10px;
+.ips--qr-expiry {
+    color: #1155cc;
+    font-weight: bold;
 }
