--- src/js/controllers/ips/ips.controller.js
@@ -1,91 +1,0 @@
-// SPDX-FileCopyrightText: Copyright (C) 2025 Opal Health Informatics Group at the Research Institute of the McGill University Health Centre <john.kildea@mcgill.ca>
-//
-// SPDX-License-Identifier: Apache-2.0
-
-import '../../../css/views/ips.view.css';
-import ipsViewers from '../../constants/ips-viewers.constants.js';
-
-(function () {
-    'use strict';
-
-    angular
-        .module('OpalApp')
-        .controller('IPSController', IPSController);
-
-    IPSController.$inject = ['$scope', '$timeout', 'Navigator', 'Params', 'ProfileSelector', 'RequestToServer'];
-
-    function IPSController($scope, $timeout, Navigator, Params, ProfileSelector, RequestToServer) {
-        const vm = this;
-
-        let navigator;
-
-        /**
-         * @description SMART Health Link data for the IPS, which usually follows # in the URL when passed to an IPS viewer.
-         */
-        let ipsLinkData;
-
-        /**
-         * @description Keeps track of the IPS viewer currently being used.
-         * @type {number}
-         */
-        let viewerIndex = 0;
-
-        vm.alertType = Params.alertTypeDanger;
-        vm.loading = false;
-        vm.loadingError = false;
-        vm.showFrame = false;
-
-        vm.ipsLink = () => ipsViewers[viewerIndex] + '#' + ipsLinkData;
-        vm.share = share;
-        vm.switchViewer = switchViewer;
-
-        activate();
-
-        function activate() {
-            vm.loading = true;
-            vm.showFrame = false;
-
-            $timeout(async () => {
-                navigator = Navigator.getNavigator();
-
-                const patient_uuid = ProfileSelector.getActiveProfile()?.patient_uuid;
-                const requestParams = Params.API.ROUTES.IPS;
-                const formattedParams = {
-                    ...requestParams,
-                    url: requestParams.url.replace('<PATIENT_UUID>', patient_uuid),
-                };
-
-                try {
-                    let result = await RequestToServer.apiRequest(formattedParams);
-                    ipsLinkData = result.data;
-                    vm.showFrame = true;
-                }
-                catch (error) {
-                    console.error(error);
-                    vm.loadingError = true;
-                }
-                finally {
-                    vm.loading = false;
-                }
-            });
-        }
-
-        /**
-         * @description Opens the IPS sharing view.
-         */
-        function share() {
-            navigator.pushPage('./views/personal/ips/ips-share.html', {
-                ipsLinkData: ipsLinkData,
-                ipsLink: vm.ipsLink(),
-            });
-        }
-
-        /**
-         * @description Cycles through which viewer is currently being used, and saves this information as the viewerIndex.
-         */
-        function switchViewer() {
-            viewerIndex = viewerIndex + 1;
-            if (viewerIndex >= ipsViewers.length) viewerIndex = 0;
-        }
-    }
-})();
