--- src/views/personal/ips/ips-share.html
@@ -1,36 +1,0 @@
-<!--
-SPDX-FileCopyrightText: Copyright (C) 2025 Opal Health Informatics Group at the Research Institute of the McGill University Health Centre <john.kildea@mcgill.ca>
-
-SPDX-License-Identifier: Apache-2.0
--->
-
-<ons-page ng-controller="IPSShareController as vm">
-    <generic-toolbar title="{{'IPS_FULL'|translate}}" style-class="fontSizeTitle"></generic-toolbar>
-
-    <network-banner></network-banner>
-
-    <div class="panel panel-default ips-outer-panel">
-        <div class="panel-heading">
-            <h3 class="panel-title">{{"SHARE"|translate}}</h3>
-        </div>
-        <div class="ips-outer-panel">
-
-            <!-- PROTOTYPE BADGE -->
-            <div class="text-badge ips-prototype-badge" style="margin-bottom: 10px">{{"IPS_PROTOTYPE"|translate}}</div>
-
-            <!-- DESCRIPTION -->
-            <p>{{"IPS_QR_CODE_DESCRIPTION"|translate}}</p>
-
-            <!-- QR CODE -->
-            <p>
-                <img class="ips-qr-code" ng-src="{{vm.qrCodeData}}" alt="{{'IPS_QR_CODE_ALT'|translate}}"/>
-            </p>
-
-            <!-- LINK LABEL -->
-            <p class="ips-link-label">{{"IPS_QR_CODE_LINK"|translate}}</p>
-
-            <!-- QR CODE LINK -->
-            <p class="ips-link" ng-click="vm.openIpsLink()">{{vm.ipsLink}}</p>
-        </div>
-    </div>
-</ons-page>
