--- src/views/personal/ips/ips.html
@@ -1,36 +1,0 @@
-<!--
-SPDX-FileCopyrightText: Copyright (C) 2025 Opal Health Informatics Group at the Research Institute of the McGill University Health Centre <john.kildea@mcgill.ca>
-
-SPDX-License-Identifier: Apache-2.0
--->
-
-<ons-page ng-controller="IPSController as vm">
-    <generic-toolbar title="{{'IPS_FULL'|translate}}" style-class="fontSizeTitle">
-        <right-content>
-            <ons-toolbar-button ng-if="!vm.loading && !vm.loadingError" ng-click="vm.share()" id="buttonShareMaterial">
-                <ons-icon icon="ion-android-share-alt" fixed-width="true"></ons-icon>
-            </ons-toolbar-button>
-        </right-content>
-    </generic-toolbar>
-
-    <network-banner></network-banner>
-
-    <!-- PROTOTYPE BADGE -->
-    <span class="text-badge ips-prototype-badge" style="margin: 10px">{{"IPS_PROTOTYPE"|translate}}</span>
-
-    <!-- BUTTON TO SWITCH VIEWER -->
-    <span class="text-badge ips-viewer-badge" ng-click="vm.switchViewer()">
-        {{"IPS_SWITCH_VIEWER"|translate}}
-    </span>
-
-    <!-- LOADING WHEEL -->
-    <loading-spinning-circle ng-show="vm.loading" loading-message="{{'LOADING'|translate}}"></loading-spinning-circle>
-
-    <!-- ERROR -->
-    <div ng-show="vm.loadingError" align="center" style="width: 95%; margin: 10px auto" ng-class="fontSizeDesc">
-        <uib-alert type="{{vm.alertType}}">{{"LOADING_ERROR"|translate}}</uib-alert>
-    </div>
-
-    <!-- IPS VIEWER -->
-    <iframe ng-if="vm.showFrame && !vm.loading" ng-src="{{vm.ipsLink()}}" title="{{'IPS_VIEWER_ALT'|translate}}" class="ips-iframe"></iframe>
-</ons-page>
